Adds abstract base class and StringField
The abstract base class describes how all Fields should look like.
A field should have:
  - position: string position in the record
  - length: string width
  - value
  - tag: (optional, incurred from standard)

The dumps method prints the value according the specs.

The loads method reads a value according to the specs; it does by using 
the parse method, which in turn calls it’s regex.

The StringField is a first generic implementation.