Adds default attribute getters / setters
Implements default getters and setters for all record field values. The
field values are exposed through the Record base class `__getattr__`.
Similarly, the values can be updated through `__setattr__`.

I was a little hesistant at first to add this functionality, but
all-in-all, it places along nice with python's default behavior and is
certainly a lot less code in comparison to explicitly defining
everything through properties.